update restriction on every used registration fee

+++ src/Fast/SisdikBundle/Controller/BiayaPendaftaranController.php
@@ -183,6 +183,12 @@ class BiayaPendaftaranController extends Controller
             throw $this->createNotFoundException('Entity BiayaPendaftaran tak ditemukan.');
         }
 
+        if ($entity->getTerpakai() === true) {
+            $this->get('session')
+                    ->setFlash('success',
+                            $this->get('translator')->trans('flash.fee.registration.update.restriction'));
+        }
+
         $editForm = $this->createForm(new BiayaPendaftaranType($this->container), $entity);
         $deleteForm = $this->createDeleteForm($id);
 
@@ -219,6 +225,14 @@ class BiayaPendaftaranController extends Controller
         if ($editForm->isValid()) {
 
             try {
+
+                if ($entity->getTerpakai() === true) {
+                    $this->get('session')
+                            ->setFlash('success',
+                                    $this->get('translator')
+                                            ->trans('flash.fee.registration.update.restriction'));
+                }
+
                 $em->persist($entity);
                 $em->flush();
 
@@ -267,18 +281,16 @@ class BiayaPendaftaranController extends Controller
                 throw $this->createNotFoundException('Entity BiayaPendaftaran tak ditemukan.');
             }
 
-            try {
-                $em->remove($entity);
-                $em->flush();
-
-                $this->get('session')
-                        ->setFlash('success',
-                                $this->get('translator')->trans('flash.fee.registration.deleted'));
-
-            } catch (DBALException $e) {
-                $message = $this->get('translator')->trans('exception.delete.restrict');
-                throw new DBALException($message);
+            if ($entity->getTerpakai() === true) {
+                $message = $this->get('translator')->trans('exception.delete.restrict.registrationfee');
+                throw new \Exception($message);
             }
+
+            $em->remove($entity);
+            $em->flush();
+
+            $this->get('session')
+                    ->setFlash('success', $this->get('translator')->trans('flash.fee.registration.deleted'));
 
         } else {
             $this->get('session')
